Handle serde attrs.
